backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Vector database and embeddings
qdrant-client>=1.7.0
sentence-transformers>=2.2.2

# HTTP client for API calls
httpx>=0.25.0

# Claude API client
anthropic>=0.7.0

# Environment variables
python-dotenv>=1.0.0
